DuckDB connection class

Implements DBIConnection.

# S4 method for duckdb_connection
dbAppendTable(conn, name, value, ..., row.names = NULL)

# S4 method for duckdb_connection
dbBegin(conn, ...)

# S4 method for duckdb_connection
dbCommit(conn, ...)

# S4 method for duckdb_connection
dbDataType(dbObj, obj, ...)

# S4 method for duckdb_connection,character
dbExistsTable(conn, name, ...)

# S4 method for duckdb_connection
dbGetInfo(dbObj, ...)

# S4 method for duckdb_connection
dbIsValid(dbObj, ...)

# S4 method for duckdb_connection,character
dbListFields(conn, name, ...)

# S4 method for duckdb_connection
dbListTables(conn, ...)

# S4 method for duckdb_connection,ANY
dbQuoteIdentifier(conn, x, ...)

# S4 method for duckdb_connection
dbQuoteLiteral(conn, x, ...)

# S4 method for duckdb_connection,character
dbRemoveTable(conn, name, ..., fail_if_missing = TRUE)

# S4 method for duckdb_connection
dbRollback(conn, ...)

# S4 method for duckdb_connection,character
dbSendQuery(conn, statement, params = NULL, ..., arrow = FALSE)

# S4 method for duckdb_connection,character,data.frame
dbWriteTable(
  conn,
  name,
  value,
  row.names = FALSE,
  overwrite = FALSE,
  append = FALSE,
  field.types = NULL,
  temporary = FALSE,
  ...
)

# S4 method for duckdb_connection
show(object)

Arguments

conn

A DBIConnection object, as returned by dbConnect().

name

The table name, passed on to dbQuoteIdentifier(). Options are:

a character string with the unquoted DBMS table name, e.g. "table_name",

a call to Id() with components to the fully qualified table name, e.g. Id(schema = "my_schema", table = "table_name")

a call to SQL() with the quoted and fully qualified table name given verbatim, e.g. SQL('"my_schema"."table_name"')

value

A data frame of values. The column names must be consistent with those in the target table in the database.

...

Other parameters passed on to methods.

row.names

Whether the row.names of the data.frame should be preserved

dbObj

A object inheriting from DBIDriver or DBIConnection

obj

An R object whose SQL type we want to determine.

statement

a character string containing SQL.

params

A list of bindings, named or unnamed.

arrow

Whether the query should be returned as an Arrow Table

overwrite

If a table with the given name already exists, should it be overwritten?

append

If a table with the given name already exists, just try to append the passed data to it

field.types

Override the auto-generated SQL types

temporary

Should the created table be temporary?

object

Any R object