duckdb: DBI Package for the DuckDB Database Management System

The DuckDB project is an embedded analytical data management system with support for the Structured Query Language (SQL). This package includes all of DuckDB and a R Database Interface (DBI) connector.

See also

Author

Maintainer: Kirill Müller kirill@cynkra.com (ORCID)

Authors:

Hannes Mühleisen hannes@cwi.nl (ORCID)

Mark Raasveldt mark.raasveldt@cwi.nl (ORCID)

Other contributors:

Stichting DuckDB Foundation [copyright holder]

Apache Software Foundation [copyright holder]

PostgreSQL Global Development Group [copyright holder]

The Regents of the University of California [copyright holder]

Cameron Desrochers [copyright holder]

Victor Zverovich [copyright holder]

RAD Game Tools [copyright holder]

Valve Software [copyright holder]

Rich Geldreich [copyright holder]

Tenacious Software LLC [copyright holder]

The RE2 Authors [copyright holder]

Google Inc. [copyright holder]

Facebook Inc. [copyright holder]

Steven G. Johnson [copyright holder]

Jiahao Chen [copyright holder]

Tony Kelman [copyright holder]

Jonas Fonseca [copyright holder]

Lukas Fittl [copyright holder]

Salvatore Sanfilippo [copyright holder]

Art.sy, Inc. [copyright holder]

Oran Agra [copyright holder]

Redis Labs, Inc. [copyright holder]

Melissa O'Neill [copyright holder]

PCG Project contributors [copyright holder]